% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{choose_custom}
\alias{choose_custom}
\title{Binomial Coefficient}
\usage{
choose_custom(n, k)
}
\arguments{
\item{n}{The number of items.}

\item{k}{The number of items to choose.}
}
\value{
The binomial coefficient (\code{n choose k}). Returns 0 if \code{k > n} or \code{k < 0}.
}
\description{
Function to calculate the binomial coefficient (\code{n choose k}).
}
\examples{
choose_custom(5, 2) # Returns 10
choose_custom(5, 6) # Returns 0
}
