% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{choose_custom_log}
\alias{choose_custom_log}
\title{Binomial Coefficient: log-gamma  form}
\usage{
choose_custom_log(n, k)
}
\arguments{
\item{n}{The number of items.}

\item{k}{The number of items to choose.}
}
\value{
The binomial coefficient (\code{n choose k}). Returns 0 if \code{k > n} or \code{k < 0}.
}
\description{
Function to calculate the binomial coefficient (\code{n choose k}) in log-gamma  form.
}
\examples{
choose_custom_log(5, 2) # Returns 10
choose_custom_log(5, 6) # Returns 0
}
