% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{dpmfHG.imperfect.group}
\alias{dpmfHG.imperfect.group}
\title{Derivative of PMF w.r.t Tx for Imperfect Group-Level Sensitivity and Specificity}
\usage{
dpmfHG.imperfect.group(ty, N, barN, Tx, b, delta, lambda, verbose = TRUE)
}
\arguments{
\item{ty}{Integer. Number of identified contaminated groups (out of \eqn{b} total groups).}

\item{N}{Integer. Total population size.}

\item{barN}{Numeric. Average size of each group.}

\item{Tx}{Integer. Number of successes (positive cases) to be distributed in the population.}

\item{b}{Integer. Total number of groups.}

\item{delta}{Numeric. Group-level false negative rate (\eqn{\delta}), the probability that a positive case is classified as negative (0 ≤ delta ≤ 1).}

\item{lambda}{Numeric. Group-level false positive rate (\eqn{\lambda}), the probability that a negative case is classified as positive (0 ≤ lambda ≤ 1).}

\item{verbose}{Logical. If TRUE, prints conditions that are not fulfilled.}
}
\value{
Numeric. The derivative of the computed PMF value w.r.t Tx adjusted for imperfect group-level sensitivity and specificity.
}
\description{
Computes derivative of the probability mass function (PMF) for group-level sensitivity and specificity,
w.r.t. Tx accounting for false negative rate (\eqn{\delta}) and false positive rate (\eqn{\lambda}).
}
\examples{
# Example usage:
pmfHG.imperfect.group(ty = 3, N = 100, barN = 4, Tx = 20, b = 4, delta = 0.7, lambda = 0.8)
}
