% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{dpmfHG.imperfect.item}
\alias{dpmfHG.imperfect.item}
\title{Derivative of PMF for Imperfect Item-Level Detection}
\usage{
dpmfHG.imperfect.item(ty, N, barN, Tx, b, delta, lambda, verbose = TRUE)
}
\arguments{
\item{ty}{Integer. Number of identified contaminated groups (out of total \code{b} groups).}

\item{N}{Integer. Total population size (base size for hypergeometric distribution).}

\item{barN}{Numeric. Average size of each group.}

\item{Tx}{Integer. Number of successes (contaminated cases to be distributed).}

\item{b}{Integer. Total number of groups.}

\item{delta}{Numeric. False negative rate (item-level probability of a true positive being missed).}

\item{lambda}{Numeric. False positive rate (item-level probability of a true negative being falsely identified as positive).}
}
\value{
Numeric. The derivative of the PMF with respect to \code{Tx}.
}
\description{
This function calculates the derivative of the probability mass function (PMF) for
imperfect item-level detection based on the provided parameters. It adjusts for
item-level sensitivity and specificity using the inclusion-exclusion principle.
}
\examples{
dpmfHG.imperfect.item(ty = 3, N = 100, barN = 10, Tx = 5, b = 10, delta = 0.1, lambda = 0.05)
}
