% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{dpmfHG.perfect}
\alias{dpmfHG.perfect}
\title{Derivative of PMF w.r.t Tx (Perfect Case)}
\usage{
dpmfHG.perfect(ty, N, barN, Tx, b, verbose = TRUE)
}
\arguments{
\item{ty}{Integer. Number of groups with observed target outcomes.}

\item{N}{Integer. Total population size.}

\item{barN}{Numeric. Average group size.}

\item{Tx}{Integer. Number of target individuals in the population.}

\item{b}{Integer. Total number of groups.}

\item{verbose}{Logical. If TRUE, prints conditions that are not fulfilled.}
}
\value{
Numeric. The derivative of the PMF for the perfect case.
}
\description{
Computes the derivative of the PMF for the perfect case as per Equation 26.
}
