% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{dqDLkGroup}
\alias{dqDLkGroup}
\title{Derivative of \eqn{q_{{\Delta\Lambda} k}} with respect to \eqn{T_x}}
\usage{
dqDLkGroup(k, N, Tx, barN, delta, lambda)
}
\arguments{
\item{k}{Integer. Number of groups considered in the summation.}

\item{N}{Integer. Total population size.}

\item{Tx}{Integer. Number of successes (positive cases) to be distributed.}

\item{barN}{Numeric. Average group size.}

\item{delta}{Numeric. False negative rate (0 ≤ delta ≤ 1).}

\item{lambda}{Numeric. False positive rate (0 ≤ lambda ≤ 1).}
}
\value{
Numeric. The derivative of \eqn{q_{\Delta\Lambda k}} with respect to \eqn{T_x}.
}
\description{
Calculates the derivative of \eqn{q_{{\Delta \Lambda} k}} with respect to \eqn{T_x} for the grouped case.
}
