% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{dqdlkItem}
\alias{dqdlkItem}
\title{Compute Derivative of \eqn{q_{\delta\lambda k} } with respect to \eqn{T_x}}
\usage{
dqdlkItem(k, N, Tx, barN, delta, lambda)
}
\arguments{
\item{k}{Integer. Number of groups considered in the summation.}

\item{N}{Integer. Total population size.}

\item{Tx}{Integer. Number of successes (positive cases) to be distributed.}

\item{barN}{Numeric. Average size of each group.}

\item{delta}{Numeric. False negative rate (item level).}

\item{lambda}{Numeric. False positive rate (item level).}
}
\value{
Numeric. The derivative of \eqn{q_{\delta\lambda k} } with respect to \eqn{T_x}.
}
\description{
Calculates the derivative of \eqn{q_{\delta\lambda k} } with respect to the
number of successes \eqn{T_x} for item-level sensitivity and specificity.
}
\examples{
dqdlk(k = 3, N = 100, Tx = 5, barN = 10, delta = 0.1, lambda = 0.05)
}
