% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN Functions.R
\name{ExpVarBN.imperfect.item}
\alias{ExpVarBN.imperfect.item}
\title{Expected Value and Variance for Grouped Binomial (Item-level Imperfect Case)}
\usage{
ExpVarBN.imperfect.item(N, Tx, barN, b, delta, lambda)
}
\arguments{
\item{N}{Numeric. The total population size.}

\item{Tx}{Numeric. The number of target individuals in the population.}

\item{barN}{Numeric. The average number of individuals per group.}

\item{b}{Numeric. The total number of groups.}

\item{delta}{Numeric. The group-level false negative rate (probability that a positive case is classified as negative).}

\item{lambda}{Numeric. The group-level false positive rate (probability that a negative case is classified as positive).}
}
\value{
A list with two components:
\itemize{
\item \code{expectation}: The expected number of detected target individuals across all groups.
\item \code{variance}: The variance of the number of detected target individuals across all groups.
}
The expectation and variance are based on the given parameters, considering the detection probabilities influenced by
the false positive and false negative rates.
}
\description{
This function calculates the expected value and variance for a grouped binomial distribution in the imperfect case,
where false positive and false negative rates are considered at item-level. The function computes the expectation and variance based
on the number of target individuals in the population, total population size, and group-level detection rates.
}
\details{
This function calculates the expectation and variance for the number of detected target individuals in a binomially distributed
scenario where imperfect detection is considered. The detection probabilities are adjusted using the false positive rate
(\code{lambda}) and the false negative rate (\code{delta}). The function calculates the expected number of detections
and the variance using the given parameters.
}
\examples{
# Example calculation for expectation and variance with group-level imperfect detection
ExpVarBN.imperfect.group(N = 1000, Tx = 10, barN = 50, b = 5, delta = 0.1, lambda = 0.05)

}
