% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{factorial_custom}
\alias{factorial_custom}
\title{Factorial Custom Function}
\usage{
factorial_custom(n)
}
\arguments{
\item{n}{A non-negative integer.}
}
\value{
The factorial of \code{n}.
}
\description{
Helper function to calculate the factorial of a non-negative integer.
}
\examples{
factorial_custom(5) # Returns 120
factorial_custom(0) # Returns 1
}
