% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN Functions.R
\name{FIpmfBN.Tx.imperfect}
\alias{FIpmfBN.Tx.imperfect}
\title{Fisher Information for Grouped Binomial (Group or Item Level Imperfect Case)}
\usage{
FIpmfBN.Tx.imperfect(
  N,
  barN,
  Tx,
  b,
  delta,
  lambda,
  method = c("AD", "ND", "PMF-SM", "PMF-HI"),
  type = c("group", "item")
)
}
\arguments{
\item{N}{Numeric. The total population size.}

\item{barN}{Numeric. The average number of individuals per group.}

\item{Tx}{Numeric. The number of target individuals in the population.}

\item{b}{Numeric. The total number of groups.}

\item{delta}{Numeric. The group or item-level false negative rate (probability that a positive case is classified as negative).}

\item{lambda}{Numeric. The group or item-level false positive rate (probability that a negative case is classified as positive).}

\item{method}{Character. Method for computing Fisher Information. One of \code{"AD"} (analytic derivative), \code{"ND"} (numerical derivative), \code{"PMF-SM"} (PMF-based approximation using Sanchez-Moreno et al (2009) paper where ty=0 to (b-1)), or \code{"PMF-HI"} (where ty=0 to b used as in Shemyakin (2023)).}

\item{type}{Character. The type of model for which the Fisher information is calculated. One of "group" (group-level model) or "item" (item-level model). Default is "group".}
}
\value{
Numeric. The Fisher information for the grouped binomial distribution in the imperfect case.
If special conditions occur, such as division by zero, the function skips those terms and continues the calculation.
}
\description{
This function calculates the Fisher information for a grouped binomial distribution in the imperfect case.
The Fisher information is calculated based on different methods (PMF-based approximation, numerical derivative, or analytical derivative)
and for either group-level or item-level models. The function computes the Fisher information over a range of possible values
for the number of target outcomes observed (ty) and returns the total Fisher information for the given parameters.
}
\details{
The Fisher information is computed based on three methods:
\itemize{
\item "PMF": PMF-based approximation method, where the Fisher information is calculated using probabilities from the PMF function.
\item "ND": Numerical derivative method, where the Fisher information is computed using finite differences of the PMF function.
\item "AD": Analytical derivative method, which uses previously defined Fisher information formulas for group and item-level models.
}
The Fisher information is accumulated over all possible values of the number of target outcomes observed (\code{ty}),
from 0 to \code{b}, where \code{b} is the total number of groups. The appropriate PMF function is used based on the \code{type} parameter
(group-level or item-level).

Special cases are handled where \code{ty} exceeds \code{Tx} and when \code{Tx} equals 0 with perfect detection rates (i.e., both \code{delta} and \code{lambda} equal 1).
The function also skips terms where the PMF is 0 to avoid division by zero errors.
}
\examples{
# Example using PMF method for group-level model
FIpmfBN.Tx.imperfect(N = 1000, barN = 50, Tx = 10, b = 5, delta = 0.1, lambda = 0.05, method = "PMF", type = "group")

# Example using numerical derivative method for item-level model
FIpmfBN.Tx.imperfect(N = 1000, barN = 50, Tx = 10, b = 5, delta = 0.1, lambda = 0.05, method = "ND", type = "item")

}
