% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN Functions.R
\name{FIpmfBN.Tx.perfect}
\alias{FIpmfBN.Tx.perfect}
\title{Fisher Information for Grouped Binomial (Perfect Case)}
\usage{
FIpmfBN.Tx.perfect(Tx, b, barN, N)
}
\arguments{
\item{Tx}{Numeric. The number of target individuals in the population.}

\item{b}{Numeric. The total number of groups.}

\item{barN}{Numeric. The average number of individuals per group.}

\item{N}{Numeric. The total population size.}
}
\value{
Numeric. The Fisher Information for the grouped binomial distribution in the perfect case.
If the parameters are invalid or special conditions occur (such as Tx = 0 or numerical instability),
the function returns 0 and prints a message indicating the issue.
}
\description{
This function calculates the Fisher information for a grouped binomial distribution in the perfect case
(when detection probability is assumed to be perfect), based on the number of target individuals in the population.
The formula for Fisher Information is derived for this scenario, and the function returns the corresponding value.
}
\details{
The Fisher Information is calculated using a formula based on the parameters of the grouped binomial model.
The function handles special cases, such as when the number of target individuals (\code{Tx}) is 0, by setting the Fisher information to 0.
It also checks for numerical instability by verifying if the denominator is close to zero and, in such cases, sets the Fisher Information to 0.
}
\examples{
# Example calculation for valid parameters
FIpmfBN.Tx.perfect(Tx = 10, b = 5, barN = 50, N = 1000)

# Example when Tx = 0
FIpmfBN.Tx.perfect(Tx = 0, b = 5, barN = 50, N = 1000)

}
