% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{FIpmfHG.Tx.imperfect.detail}
\alias{FIpmfHG.Tx.imperfect.detail}
\title{Fisher Information for Imperfect Detection}
\usage{
FIpmfHG.Tx.imperfect.detail(
  N,
  barN,
  Tx,
  b,
  delta,
  lambda,
  method = c("AD", "ND", "PMF-HI", "PMF-SM"),
  type = c("group", "item"),
  verbose = TRUE
)
}
\arguments{
\item{N}{Integer. Total population size (base size for hypergeometric distribution).}

\item{barN}{Numeric. Average size of each group.}

\item{Tx}{Integer. Number of successes (contaminated cases to be distributed).}

\item{b}{Integer. Total number of groups.}

\item{delta}{Numeric. False negative rate.}

\item{lambda}{Numeric. False positive rate.}

\item{method}{Character. Method for computing Fisher Information. One of \code{"AD"} (analytic derivative), \code{"ND"} (numerical derivative), \code{"PMF-SM"} (PMF-based approximation using Sanchez-Moreno et al (2009) paper where ty=0 to (b-1)), or \code{"PMF-HI"} (where ty=0 to b used as in Shemyakin (2023)).}

\item{type}{Character. Detection type. One of \code{"group"} or \code{"item"}.}

\item{verbose}{Logical. If TRUE, prints conditions that are not fulfilled.}
}
\value{
Numeric. The calculated Fisher Information.
}
\description{
This function calculates the Fisher Information for imperfect detection using different
methods (analytic, numerical, or PMF-based approximation) and detection types (group or item level).
This function also provides probability, derivative of ty w.r.t Tx and finally Fisher information for a given Tx
}
\examples{
FIpmfHG.Tx.imperfect(N = 100, barN = 4, Tx = 20, b = 4, delta = 0.7, lambda = 0.8, method = "AD", type = "item")
}
