% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{pmf_hg_group_imperfect}
\alias{pmf_hg_group_imperfect}
\title{PMF for Imperfect Group-Level Sensitivity and Specificity}
\usage{
pmf_hg_group_imperfect(ty, N, barN, Tx, b, delta, lambda, verbose = TRUE)
}
\arguments{
\item{ty}{Integer. Number of identified contaminated groups (out of \eqn{b} total groups).}

\item{N}{Integer. Total population size.}

\item{barN}{Numeric. Average size of each group.}

\item{Tx}{Integer. Number of successes (positive cases) to be distributed in the population.}

\item{b}{Integer. Total number of groups.}

\item{delta}{Numeric. Group-level false negative rate (\eqn{\Delta}), the probability that a positive case is classified as negative (0 ≤ delta ≤ 1).}

\item{lambda}{Numeric. Group-level false positive rate (\eqn{\Lambda}), the probability that a negative case is classified as positive (0 ≤ lambda ≤ 1).}

\item{verbose}{Logical. If TRUE, prints conditions that are not fulfilled.}
}
\value{
Numeric. The computed PMF value adjusted for imperfect group-level sensitivity and specificity.
}
\description{
Computes the probability mass function (PMF) for group-level sensitivity and specificity,
accounting for false negative rate (\eqn{\Delta}) and false positive rate (\eqn{\Lambda}).
}
\examples{
# Example usage:
pmf_hg_group_imperfect(ty = 3, N = 100, barN = 4, Tx = 20, b = 4, delta = 0.7, lambda = 0.8)
}
