% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{pmf_hg_perfect}
\alias{pmf_hg_perfect}
\title{PMF for Grouped Hypergeometric Distribution under Perfect Test}
\usage{
pmf_hg_perfect(ty, N, barN, Tx, b)
}
\arguments{
\item{ty}{Number of groups where the target outcome is observed.}

\item{N}{Total population size.}

\item{barN}{Average number of individuals per group.}

\item{Tx}{Number of target individuals in the population.}

\item{b}{Total number of groups.}
}
\value{
The probability mass function (PMF) value for the given parameters.
}
\description{
Function to calculate the PMF for a grouped hypergeometric distribution under a perfect test.
}
\examples{
pmf_hg_perfect(ty = 2, N = 100, barN = 4, Tx = 20, b = 4)
}
