% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{pmfHG.imperfect.item}
\alias{pmfHG.imperfect.item}
\title{Probability Mass Function (PMF) for Imperfect Item-Level Sensitivity and Specificity}
\usage{
pmfHG.imperfect.item(ty, N, barN, Tx, b, delta, lambda, verbose = TRUE)
}
\arguments{
\item{ty}{Integer. Number of identified contaminated groups (out of total \eqn{b} groups).}

\item{N}{Integer. Total population size.}

\item{barN}{Numeric. Average size of each group.}

\item{Tx}{Integer. Number of successes (contaminated cases to be distributed).}

\item{b}{Integer. Total number of groups.}

\item{delta}{Numeric. False negative rate (item level probability of a true positive being missed).}

\item{lambda}{Numeric. False positive rate (item level probability of a true negative being falsely identified as positive).}

\item{verbose}{Logical. If TRUE, prints conditions that are not fulfilled.}
}
\value{
Numeric. The computed probability mass function value.
}
\description{
Calculates the probability mass function (PMF) with item-level sensitivity and
specificity, based on the number of identified contaminated groups.
}
\examples{
pmfHG.imperfect.item(ty = 2, N = 100, barN = 10, Tx = 5, b = 5, delta = 0.1, lambda = 0.05)
}
