% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{qDLkGroup}
\alias{qDLkGroup}
\title{Compute \eqn{q_{{\Delta\Lambda} k}} for group-Level Sensitivity and Specificity}
\usage{
qDLkGroup(k, N, Tx, barN, delta, lambda)
}
\arguments{
\item{k}{Integer. Number of groups considered in the summation.}

\item{N}{Integer. Total population size.}

\item{Tx}{Integer. Number of successes (positive cases) to be distributed.}

\item{barN}{Numeric. Average group size.}

\item{delta}{Numeric. False negative rate (0 ≤ delta ≤ 1).}

\item{lambda}{Numeric. False positive rate (0 ≤ lambda ≤ 1).}
}
\value{
Numeric. The probability \eqn{q_{\Delta\Lambda k}}.
}
\description{
Calculates \eqn{q_{{\Delta\Lambda} k}}, the probability that none are identified as contaminated for the grouped case.
}
