% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{qdlkItem}
\alias{qdlkItem}
\title{Compute \eqn{q_{\delta \lambda k} } for Item-Level Sensitivity and Specificity}
\usage{
qdlkItem(k, N, Tx, barN, delta, lambda)
}
\arguments{
\item{k}{Integer. Number of groups considered in the summation.}

\item{N}{Integer. Total population size.}

\item{Tx}{Integer. Number of successes (positive cases) to be distributed.}

\item{barN}{Numeric. Average size of each group.}

\item{delta}{Numeric. False negative rate (item level).}

\item{lambda}{Numeric. False positive rate (item level).}
}
\value{
Numeric. The computed value of \eqn{q_{\delta\lambda k} }.
}
\description{
Calculates the probability \eqn{q_{\delta \lambda k} } that none of the items
in the groups are identified as contaminated, considering item-level false
negative and false positive rates.
}
\examples{
qdlkqdlkItem(k = 3, N = 100, Tx = 5, barN = 10, delta = 0.1, lambda = 0.05)
}
