% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG Functions.R
\name{safe_digamma}
\alias{safe_digamma}
\title{Safe Digamma Function}
\usage{
safe_digamma(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
The digamma of \code{x} if \code{x > 0}, otherwise 0.
}
\description{
A safe version of the digamma function that returns 0 for invalid arguments (\code{x <= 0}).
}
\examples{
safe_digamma(5) # Returns digamma(5)
safe_digamma(-1) # Returns 0
}
