% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BN Functions.R
\name{simGroupedBN}
\alias{simGroupedBN}
\title{Simulate Grouped Binomial Sampling with Sensitivity and Specificity Adjustments}
\usage{
simGroupedBN(N, barN, Tx, b, delta, lambda, NoSim, type = c("item", "group"))
}
\arguments{
\item{N}{Integer. The total population size.}

\item{barN}{Integer. The group size.}

\item{Tx}{Integer. The number of infectives in the population.}

\item{b}{Integer. The number of group samples.}

\item{delta}{Numeric (0 to 1). Sensitivity of detection (delta = 1 removes sensitivity issues).}

\item{lambda}{Numeric (0 to 1). Specificity of detection (lambda = 1 removes specificity issues).}

\item{NoSim}{Integer. The number of simulation iterations.}

\item{type}{Character. Either \code{"item"} or \code{"group"}, specifying the detection model.}
}
\value{
A list containing simulation results:
\itemize{
\item If \code{type = "item"}:
\itemize{
\item \code{tydat}: Matrix of total detections.
\item \code{tydatdelta}: Matrix of detections considering sensitivity.
\item \code{tydatlambda}: Matrix of detections considering specificity.
\item \code{tydatdeltalambda}: Matrix of detections considering both sensitivity and specificity.
}
\item If \code{type = "group"}:
\itemize{
\item \code{tydat}: Matrix of total detections.
\item \code{tydatDelta}: Matrix of detections considering sensitivity.
\item \code{tydatLambda}: Matrix of detections considering specificity.
\item \code{tydatDeltaLambda}: Matrix of detections considering both sensitivity and specificity.
}
}
}
\description{
This function simulates a grouped binomial sampling process, accounting for imperfect sensitivity and specificity in detection.
}
\examples{
simGroupedBN(N = 1000, barN = 4, Tx = 20, b = 4, delta = 0.7, lambda = 0.8, NoSim = 100, type = "item")

}
