% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_numbers.R
\name{subtract_numbers}
\alias{subtract_numbers}
\title{Subtract Two Numbers}
\usage{
subtract_numbers(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The difference of x and y.
}
\description{
Subtract Two Numbers
}
